pyproject.toml
[tool.poetry]
name = "crawlerservice-py"
version = "0.1.0"
description = ""
authors = ["Moein Moeinnia <Moeinnia@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
psycopg2 = "^2.9.9"
langchain = "^0.1.11"
confluent-kafka = "^2.3.0"
bs4 = "^0.0.2"
langchain-community = "^0.0.27"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.1"
unstructured = { version = "^0.14.7",extras = ["all-docs"] }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
